adds component for WaitList, removes Reservation from RetailerView
