+++ src/components/WaitList/index.js
@@ -1,0 +1,2 @@
+import WaitList from "../WaitingList";
+export default WaitList;
