+++ src/views/RetailerView/RetailerView.module.scss
@@ -7,54 +7,6 @@
   padding-top: 5%;
   overflow: hidden;
 
-  // both headers style, waitlist and holdlist
-
-  .header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    h3 {
-      font-size: 2em;
-      font-weight: 400;
-    }
-
-    .time {
-      display: flex;
-      justify-content: flex-start;
-      align-items: center;
-
-      img {
-        height: 36px;
-        margin-right: 6px;
-      }
-    }
-  }
-
-  .WaitList {
-    height: 100%;
-    padding: 0 12px;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    display: flex;
-    flex-direction: column;
-
-    .listContainer {
-      overflow-y: scroll;
-      padding-bottom: 35vh;
-
-      &::-webkit-scrollbar {
-        display: none;
-      }
-
-      -ms-overflow-style: none; /* IE and Edge */
-      scrollbar-width: none; /* Firefox */
-    }
-  }
-
   .HoldList {
     position: absolute;
     background: $darkgreen;
@@ -113,18 +65,6 @@
           transition: opacity 0.25s;
         }
       }
-    }
-
-    .listContainer {
-      overflow: scroll;
-      padding-bottom: 35vh;
-
-      &::-webkit-scrollbar {
-        display: none;
-      }
-
-      -ms-overflow-style: none; /* IE and Edge */
-      scrollbar-width: none; /* Firefox */
     }
   }
 }
